had to cast fcst to int after divide by 10000

+++ scripts/matsMetaDataForApps/createMetaData/mysql/metexpress/MEanomalycor.py
@@ -228,7 +228,7 @@ class MEAnomalycor:
                         fcst = int(list(line2.values())[0])
                         temp_fcsts_orig.add(fcst)
                         if fcst % 10000 == 0:
-                            fcst = fcst / 10000
+                            fcst = int(fcst / 10000)
                         temp_fcsts.add(fcst)
 
                     get_fcsts_late = "select distinct fcst_lead from \
@@ -240,7 +240,7 @@ class MEAnomalycor:
                         fcst = int(list(line2.values())[0])
                         temp_fcsts_orig.add(fcst)
                         if fcst % 10000 == 0:
-                            fcst = fcst / 10000
+                            fcst = int(fcst / 10000)
                         temp_fcsts.add(fcst)
 
                     per_mvdb[mvdb][model]['fcsts'] = list(map(str,sorted(temp_fcsts)))
